requirements/project.txt
django-taggit
django-tinymce
django-reversion==1.5.3
